requirements.txt
Django==1.7
django-absolute==0.3
django-profiles==0.2
-e git+https://github.com/hacklabr/django-recaptcha.git@bbfbf843055596e9d3eeef161af7e57335dda505#egg=django-recaptcha-dev
-e git+https://github.com/hacklabr/python-openid.git@5a0f618873d4a0bd757ccb48df4ee7145acd130c#egg=python_openid-dev
django-braces==1.3.1
django-allauth==0.18.0
